+++ R/physignal.r
@@ -10,8 +10,11 @@
 #'   in a dataset relative to what is expected under a Brownian motion model of evolution. For geometric
 #'   morphometric data, the approach is a mathematical generalization of the Kappa statistic (Blomberg et al. 
 #'   2003) appropriate for highly multivariate data (see Adams 2014).Significance testing 
-#'   is found by permuting the shape data among the tips of the phylogeny. Note that this 
-#'   method can be quite slow as ancestral states must be estimated for every iteration.
+#'   is found by permuting the shape data among the tips of the phylogeny. In addition, a multivariate 
+#'   effect size describing the strength of the effect is 
+#'   estimated from the empirically-generated sampling distribution (see details in Adams and Collyer 2019).
+#'   Values from these distributions are log-transformed prior to effect size estimation, 
+#'   to assure normally distributed data. 
 #' 
 #' This function can also be used with univariate data (i.e. centroid size) if imported as matrix with rownames
 #' giving the taxa names. In this case, the estimate of phylogenetic signal is identical to that found using the 
@@ -41,6 +44,7 @@
 #' @return Function returns a list with the following components: 
 #'   \item{phy.signal}{The estimate of phylogenetic signal}
 #'   \item{pvalue}{The significance level of the observed signal}
+#'   \item{pvalue}{The significance level of the observed signal}
 #'   \item{random.K}{Each random K-statistic from random permutations}
 #'   \item{permutations}{The number of random permutations used in the resampling procedure}
 #'   \item{call}{The matched call}
@@ -48,6 +52,9 @@
 #' data: behavioral traits are more labile. Evolution, 57:717-745.
 #' @references Adams, D.C. 2014. A generalized K statistic for estimating phylogenetic signal from shape and 
 #' other high-dimensional multivariate data. Systematic Biology.  63:685-697.
+#' @references Adams, D.C. and M.L. Collyer. 2019. Comparing the strength of modular signal, and evaluating 
+#' alternative modular hypotheses, using covariance ratio effect sizes with morphometric data. 
+#' Evolution. 73:2352-2367.
 #' @examples
 #' data(plethspecies) 
 #' Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment    
@@ -114,7 +121,8 @@ physignal<-function(A,phy,iter=999, seed=NULL, print.progress=TRUE){
   } else K.rand <- sapply(1:(iter+1), 
                           function(j) Kmult(as.matrix(x.rand[[j]]), invC,D.mat, ones, a.adj))
   p.val <- pval(K.rand)
-  out <- list(phy.signal=K.rand[[1]], pvalue=p.val, random.K = K.rand,
+  Z <- effect.size(log(K.rand), center=TRUE) 
+  out <- list(phy.signal=K.rand[[1]], pvalue=p.val, Z = Z, random.K = K.rand,
               permutations = iter+1, call=match.call())
   class(out) <- "physignal"
   out