+++ NEWS.md
@@ -1,4 +1,8 @@
-# CHANGES IN GEOMORPH 3.1.4 (Patch Release)
+# CHANGES IN GEOMORPH 3.2.0 (Minor Release)
+
+### New Features
+* Effect sizes provided in the output of `compare.evol.rates`, `compare.multi.evol.rates`,`integration.test`,
+`modularity.test`, `phylo.integration`, `phylo.modularity`, `physignal`, `two.b.pls`
 
 ### OTHER CHANGES
 * Adjustment in 'class' statements to be compliant with new handling of objects in R 4.0